requirements.txt
beautifulsoup4==4.12.3
click==8.1.7
jsonschema==4.23.0
requests==2.32.3
urllib3==2.2.3
selenium==4.27.1
PyYAML==6.0.2
